+++ nfe_rhea/dialogs.hpp
@@ -1,0 +1,9 @@
+#include "dialogs\defines.hpp"
+
+#include "dialogs\login.hpp"
+#include "dialogs\main.hpp"
+#include "dialogs\conditions.hpp"
+#include "dialogs\input.hpp"
+#include "dialogs\playerMenu.hpp"
+#include "dialogs\taskMenu.hpp"
+#include "dialogs\debug.hpp"
