+++ nfe_rhea/dialogs/defines.hpp
@@ -1,0 +1,10 @@
+#include "\A3\ui_f\hpp\defineResincl.inc"
+#include "\A3\ui_f\hpp\defineCommonColors.inc"
+#include "\A3\ui_f\hpp\defineDIKCodes.inc"
+
+#define GUI_GRID_X	(safezoneX)
+#define GUI_GRID_Y	(safezoneY)
+#define GUI_GRID_W	(safezoneW / 40)
+#define GUI_GRID_H	(safezoneH / 25)
+#define GUI_GRID_WAbs	(safezoneW)
+#define GUI_GRID_HAbs	(safezoneH)
